feat(examples): use siftsmalllearn to train faiss indexer
