+++ faiss-search/app.py
@@ -9,6 +9,7 @@ import string
 import random
 import numpy as np
 
+from fvecs_read import fvecs_read
 from jina.flow import Flow
 
 RANDOM_SEED = 14
@@ -24,29 +25,12 @@ def get_random_ws(workspace_path, length=8):
     return os.path.join(workspace_path, dn)
 
 
-def fvecs_read(filename: str, c_contiguous: bool = True) -> np.ndarray:
-    fv = np.fromfile(filename, dtype=np.float32)
-    if fv.size == 0:
-        return np.zeros((0, 0))
-    dim = fv.view(np.int32)[0]
-    assert dim > 0
-    fv = fv.reshape(-1, 1 + dim)
-    if not all(fv.view(np.int32)[:, 0] == dim):
-        raise IOError("Non-uniform vector sizes in " + filename)
-    fv = fv[:, 1:]
-    if c_contiguous:
-        fv = fv.copy()
-    return fv
-
-
-def read_data(db_file_path: str, max_sample_size: int = -1) -> Generator[bytes, Any, Any]:
+def read_data(db_file_path: str) -> Generator[bytes, Any, Any]:
     vectors = fvecs_read(db_file_path)
     num_vectors = vectors.shape[0]
-    batch_size = 1 if max_sample_size == -1 else max_sample_size
-    num_batches = int(num_vectors / batch_size)
-    for i in range(1, num_batches + 1):
-        start_batch = (i - 1) * batch_size
-        end_batch = i * batch_size if i * batch_size < num_vectors else num_vectors
+    for i in range(1, num_vectors + 1):
+        start_batch = (i - 1)
+        end_batch = i if i < num_vectors else num_vectors
         yield vectors[start_batch: end_batch].tobytes()
 
 
@@ -72,21 +56,21 @@ def save_topk(resp, output_fn=None):
 
 @click.command()
 @click.option('--task', '-t')
-@click.option('--num_docs', '-n', default=50)
+@click.option('--batch_size', '-n', default=50)
 @click.option('--top_k', '-k', default=5)
-def main(task, num_docs, top_k):
+def main(task, batch_size, top_k):
     os.environ['TMP_WORKSPACE'] = get_random_ws(os.environ['TMP_DATA_DIR'])
     if task == 'index':
         data_path = os.path.join(os.environ['TMP_DATA_DIR'], 'siftsmall_base.fvecs')
         flow = Flow().load_config('flow-index.yml')
         with flow.build() as fl:
-            fl.index(read_data(data_path, num_docs), batch_size=num_docs)
+            fl.index(read_data(data_path), batch_size=batch_size)
     elif task == 'query':
         data_path = os.path.join(os.environ['TMP_DATA_DIR'], 'siftsmall_query.fvecs')
         flow = Flow().load_config('flow-query.yml')
         with flow.build() as fl:
             ppr = lambda x: save_topk(x, os.path.join(os.environ['TMP_DATA_DIR'], 'query_results.txt'))
-            fl.search(buffer=read_data(data_path, num_docs), callback=ppr, top_k=top_k)
+            fl.search(buffer=read_data(data_path), callback=ppr, top_k=top_k)
     else:
         raise NotImplementedError(
             f'unknown task: {task}. A valid task is either `index` or `query`.')
